refactor!: remove usePhone hook, change hideSpaceAfterDialCode with charAfterDialCode
BREAKING CHANGE: `usePhone` hook has been removed

BREAKING CHANGE: `hideSpaceAfterDialCode` prop was changed with `charAfterDialCode`
