+++ packages/docs/docs/02-Usage/01-PhoneInput.md
@@ -43,7 +43,7 @@ inputProps={{ autoFocus: true }}
 | disabled                        | `boolean`             | Disable phone input and country selector.                                                                                                                                                                            | `false`                     |
 | prefix                          | `string`              | Prefix for phone value.                                                                                                                                                                                              | `"+"`                       |
 | defaultMask                     | `string`              | This mask will apply on countries that does not have specified mask.                                                                                                                                                 | `"............"` (12 chars) |
-| hideSpaceAfterDialCode          | `boolean`             | Hide space after country dial code                                                                                                                                                                                   | `false`                     |
+| charAfterDialCode               | `string`              | Char that renders after country dial code.                                                                                                                                                                           | ` `                         |
 | historySaveDebounceMS           | `number`              | Save value to history if there were not any changes in provided milliseconds timeslot.<br />Undo/redo (ctrl+z/ctrl+shift+z) works only with values that are saved in history                                         | `200`                       |
 | disableCountryGuess             | `boolean`             | Disable country guess on value change.<br />- _onCountryGuess_ callback would not be called.                                                                                                                         | `false`                     |
 | disableDialCodePrefill          | `boolean`             | Disable dial code prefill on initialization.<br />Dial code prefill works only when "empty" phone value have been provided.                                                                                          | `false`                     |
