+++ packages/docs/docs/04-Advanced Usage/01-usePhoneInput.md
@@ -28,7 +28,7 @@ const { country, setCountry, phone, handlePhoneValueChange, inputRef } =
 | countries                       | `CountryData[]`           | An array of available countries to select (and guess).                                                                                                                                                               | `defaultCountries`          |
 | prefix                          | `string`                  | Prefix for phone value.                                                                                                                                                                                              | `"+"`                       |
 | defaultMask                     | `string`                  | This mask will apply on countries that does not have specified mask.                                                                                                                                                 | `"............"` (12 chars) |
-| hideSpaceAfterDialCode          | `boolean`                 | Hide space after country dial code.                                                                                                                                                                                  | `false`                     |
+| charAfterDialCode               | `string`                  | Char that renders after country dial code.                                                                                                                                                                           | ` `                         |
 | historySaveDebounceMS           | `number`                  | Save value to history if there were not any changes in provided milliseconds timeslot.<br />Undo/redo (ctrl+z/ctrl+shif+z) works only with values that are saved in history.                                         | `200`                       |
 | disableCountryGuess             | `boolean`                 | Disable country guess on value change.<br />- _onCountryGuess_ callback would not be called.                                                                                                                         | `false`                     |
 | disableDialCodePrefill          | `boolean`                 | Disable dial code prefill on initialization.<br />Dial code prefill works only when "empty" phone value have been provided.                                                                                          | `false`                     |
