+++ src/index.ts
@@ -5,8 +5,6 @@ export { FlagEmoji } from './components/FlagEmoji/FlagEmoji';
 export type { PhoneInputProps } from './components/PhoneInput/PhoneInput';
 export { PhoneInput } from './components/PhoneInput/PhoneInput';
 export { defaultCountries } from './data/countryData';
-export type { UsePhoneConfig } from './hooks/usePhone';
-export { usePhone } from './hooks/usePhone';
 export { usePhoneInput } from './hooks/usePhoneInput';
 export { usePhoneValidation } from './hooks/usePhoneValidation';
 export type { CountryIso2 } from './types';
